refactor: LV3 입력 관련 수정 역활과 구현의 분리 객체 지향적 설계 시도 중
