+++ src/lv03/Input/CalculatorInput.java
@@ -1,0 +1,16 @@
+package lv03.Input;
+
+import java.util.Scanner;
+
+public class CalculatorInput implements Input {
+    private final Scanner scanner = new Scanner(System.in);
+
+    public String input() {
+        System.out.print(">> ");
+        return scanner.nextLine();
+    }
+
+    public void end() {
+        scanner.close();
+    }
+}
