+++ src/lv03/Input/Input.java
@@ -1,0 +1,8 @@
+package lv03.Input;
+
+import java.util.Scanner;
+
+public interface Input {
+    String input();
+    void end();
+}
