+++ src/lv03/calculations/Calculation.java
@@ -1,4 +1,4 @@
-package lv03.calculation;
+package lv03.calculations;
 
 public interface Calculation {
     public double calculate(double num1, double num2);
