+++ src/lv03/calculations/SquareCalculation.java
@@ -1,4 +1,4 @@
-package lv03.calculation;
+package lv03.calculations;
 
 public class SquareCalculation implements Calculation {
     @Override
