+++ src/lv03/menus/MainMenu.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.menus;
 
 public class MainMenu extends Menu {
     private boolean state = true;
