+++ src/lv03/menus/Menu.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.menus;
 
 public abstract class Menu {
     boolean state = false;
